requirements.txt
transformers
pillow
tesseract
pytesseract
imutils
opencv-python
scipy
torch
torchvision
streamlit